+++ docs/changelog.rst
@@ -321,6 +321,8 @@ Detailed list of changes
 
 - edit-in-kitty: harden the code used to parse messages from clients
 
+- Linux: :option:`kitty --single-instance`: Only accept commands from processes running as the same user
+
 - Clipboard protocol: Report an ``EFBIG`` error to programs that try to write
   more data to the clipboard than allowed by :opt:`clipboard_max_size`, instead
   of silently truncating their data. Also fix :opt:`clipboard_max_size` being